update Matlab scripts

+++ src/Matlab/AnalysisABC_comp.m
@@ -1,22 +1,48 @@
 %% Analysis of GT data using Foll's WFABC_1 aad WFABC_2 programs
 %
 %%
+% Run the script in MATLAB (press F5). The code will prompt the user for
+% any required action. The output of this code is estimates of selection
+% coefficients stored in the file:
+% ABC_medium_simple_old_collected_extended_Tend1000.csv 
 
 % this code :
-%      1. loads data (in .mat or .dat format)
+%      1. loads data in .dat format
 %      2. run WFABC_1 and WFABC_2 
 %      3. calculate the ML estimate and save
 % Last updated 26-Nov 2017
 % Author: M Saqib Sohail
-
-
+%
+% Last updated: 16-July 2019
+%               -automated assignment of various variable
+%               -estimates saved to .csv file
+%%
 clc
 clear all
 close all
 
 
-thisSet = 'medium_simple';%'medium_complex';
-numItr = 1%100;
+repeatInput = 1;
+disp('-------------------------------------------------------------------')
+disp(' ')
+disp(' This code will run WFABCv1.1 script on ground truth data. ')
+disp(' ')
+disp('Dataset to analyze: (1) Medium simple, (2) Medium complex');
+prompt = ' ';
+while(repeatInput == 1)
+    str = input(prompt,'s');
+    if(length(str) == 1 && str == '1')
+        repeatInput = 0;
+        thisSet = 'medium_simple';
+    elseif(length(str) == 1 && str == '2')
+        repeatInput = 0;
+        thisSet = 'medium_complex';
+    else
+        disp('Unexpected input, please pren ''1'' or ''2''. Dataset to analyze: (1) Medium simple, (2) Medium complex');
+        prompt = ' ';
+    end
+end
+numItr = 100;
 useSameT = 0; % flag that controls if 
               %     1: usable T will be loaded from data
               %     0: supplied by user in variable Tuse
@@ -35,10 +61,10 @@ timeToStartWFABC1 = zeros(1, numItr);
 timeWFABC1 = zeros(1, numItr);
 timeWFABC2 = zeros(1, numItr);
 timeWholeCodeFoll = zeros(1, numItr);
-
+errorExit = 1;
 for thisItr = 1:numItr
     tic
-%%  1. Load data (in .mat or .dat format)
+%%  1. Load data in .mat or .dat format)
     
     thisItr
     dirNameScriptFile = pwd;    
@@ -51,9 +77,9 @@ for thisItr = 1:numItr
         pause
     end
     
+    makeDirNames();
     [dirNameData, dirNameAnalysis, dirNameABC] = loadDirNames(fileNameContainingDirPath);
     dirNameRandPermFiles = [dirNameData(1:end-5) chosenSlash 'RandPermFiles' chosenSlash];
-    dirNameData = [dirNameData thisSet chosenSlash];
     dirNameAnalysis = [dirNameAnalysis thisSet chosenSlash];
 
     
@@ -67,32 +93,96 @@ for thisItr = 1:numItr
         mkdir(dirNameAnalysisFoll)        
     end
     
-    display('Loading file...') 
+    % check if output .csv file already exists
+    if(thisItr == 1)
+        dirNameSaveCSVFile = [];
+        if(strcmp(thisSet, 'medium_simple'))
+            fileNamePaperDataCompABC = ['ABC_medium_simple_old_collected_extended_Tend' num2str(T) '.csv'];
+        elseif(strcmp(thisSet, 'medium_complex'))
+            fileNamePaperDataCompABC = ['ABC_medium_complex_old_collected_extended_Tend' num2str(T) '.csv'];
+        end
+
+        if(exist([dirNameSaveCSVFile fileNamePaperDataCompABC], 'file') == 2)
+            disp(['The following file already exists ' dirNameSaveCSVFile fileNamePaperDataCompABC])
+            prompt = 'Overwrite? (y) yes, (n) no. Your choince: ';
+            repeatInput2 = 1;
+            while(repeatInput2 == 1)
+                str = input(prompt,'s');
+                if(length(str) == 1 && str == 'y')
+                    repeatInput2 = 0;
+
+                elseif(length(str) == 1 && str == 'n')
+                    repeatInput2 = 0;
+                    disp('WFABC not run. Rename existing file and run this code again.')
+                    break
+                else
+                    disp('Unexpected input, please pren ''y'' or ''n''. Overwrite? (y) yes, (n) no. Your choince: ');
+                    prompt = ' ';
+                end
+            end        
+        end
+    end
+    
+    % load data to process
+    disp('Loading file...') 
     
     fileName = ['wfsim_' thisSet '_' num2str(thisItr-1) '_T' num2str(T) '_ns' num2str(ng) '_dt' num2str(dT) '.dat'];
     
     fileNameFoll = ['Foll_' fileName(1:end-4) '.txt'];
     fileNameFollFlipVec = [fileNameFoll(1:end-4) '_flip.dat'];
+    if(exist([dirNameDataFoll fileNameFollFlipVec], 'file') ~= 2)
+        disp('Error: data file not found.')
+        disp(' ')
+        disp('Possible cause: Run DataGenWFABC_comp.m prior to running this Analysis script.')
+        break
+    end
     flipVec = dlmread([dirNameDataFoll fileNameFollFlipVec]);
     dirNameDataFollLinux = checkDirName(dirNameDataFoll);
     dirNameAnalysisFollLinux = checkDirName(dirNameAnalysisFoll);
     
     mainDir = pwd;
+    if(exist(dirNameABC, 'dir') ~= 7)
+        disp('Download WFABC from http://jjensenlab.org/software and instal it in directory')
+        disp('.../paper-MPL-inference-master/src/external/')
+        disp('Verify the installation directory name matches the one given in file dirNames.txt line 12')
+        break
+    end
     
     % copy the WFABC source files into the data directory. From the data
     % directory, we will move the analysis files to the analysis directory.
     if(thisItr == 1)
-        sourceDirName = dirNameABC;%'/local/staff/ee/mssohail/Matlab Codes/H3N2 evolution/WFABC_v1.1/WFABC_v1.1/binaries/Linux/';
+        sourceDirName = dirNameABC;
         sourceDirNameLinux = checkDirName(sourceDirName);
-
         destDirNameLinux = checkDirName(dirNameDataFoll);
+        
+        sourceDirNameWindows_hasSpace = checkDirNameWindows(sourceDirName);
+        destDirNameWindows_hasSpace = checkDirNameWindows(dirNameDataFoll);
+        
+        if(ispc)
+            if(sourceDirNameWindows_hasSpace == true || destDirNameWindows_hasSpace == true)
+                disp('Error: Directory names can not have space in them. Check:')
+                disp(sourceDirName)
+                disp(dirNameDataFoll)
+                break
+            end
+        end
+        
         wfabcFile1 = 'wfabc_1';
         wfabcFile2 = 'wfabc_2';
-        commandCopyFiles1 = ['cp ' sourceDirNameLinux wfabcFile1 ' ' destDirNameLinux];
+        if(ispc)
+            commandCopyFiles1 = ['copy ' sourceDirName wfabcFile1 '.exe ' dirNameDataFoll(1:end-1)];
+        else
+            commandCopyFiles1 = ['cp ' sourceDirNameLinux wfabcFile1 ' ' destDirNameLinux];
+        end
+        
         [statusCopyFiles1, cmdoutCopyFiles1] = system(commandCopyFiles1);
         disp(cmdoutCopyFiles1)
 
-        commandCopyFiles2 = ['cp ' sourceDirNameLinux chosenSlash wfabcFile2 ' ' destDirNameLinux];
+        if(ispc)
+            commandCopyFiles2 = ['copy ' sourceDirName wfabcFile2 '.exe ' dirNameDataFoll];
+        else
+            commandCopyFiles2 = ['cp ' sourceDirNameLinux chosenSlash wfabcFile2 ' ' destDirNameLinux];
+        end
         [statusCopyFiles2, cmdoutCopyFiles2] = system(commandCopyFiles2);
         disp(cmdoutCopyFiles2)
     end
@@ -102,7 +192,12 @@ for thisItr = 1:numItr
     %% 2. run WFABC_1 and WFABC_2 
     
     % run WFABC_1 command of Foll
-    command1 = ['./wfabc_1 ' fileNameFoll];
+    if(ispc)
+        command1 = ['wfabc_1 ' fileNameFoll];
+    else
+        command1 = ['./wfabc_1 ' fileNameFoll];
+    end
+    
     timeToStartWFABC1(thisItr) = toc;
     [status1, cmdout1] = system(command1);
     timeWFABC1(thisItr) = toc;
@@ -110,7 +205,12 @@ for thisItr = 1:numItr
     
     
     % run WFABC_2 command of Foll
-    command2 = ['./wfabc_2 -fixed_N ' num2str(Nin) ' -ploidy 1 -min_s -1 -max_s 0.5 ' fileNameFoll];
+    if(ispc)
+        command2 = ['wfabc_2 -fixed_N ' num2str(Nin) ' -ploidy 1 -min_s -1 -max_s 0.5 ' fileNameFoll];
+    else
+        command2 = ['./wfabc_2 -fixed_N ' num2str(Nin) ' -ploidy 1 -min_s -1 -max_s 0.5 ' fileNameFoll];
+    end
+    
     
     [status2, cmdout2] = system(command2);
     timeWFABC2(thisItr) = toc;
@@ -120,31 +220,48 @@ for thisItr = 1:numItr
     
     % copy all 4 analysis files to the analysis folder
     if(exist([fileNameFoll(1:end-4) '_Ne_bootstrap.txt'], 'file') == 2)
-        commandCopyAnalysisFile1 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_Ne_bootstrap.txt ' dirNameAnalysisFollLinux];
+        if(ispc)
+            commandCopyAnalysisFile1 = ['move ' dirNameDataFoll fileNameFoll(1:end-4) '_Ne_bootstrap.txt ' dirNameAnalysisFoll(1:end-1)];
+        else
+            commandCopyAnalysisFile1 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_Ne_bootstrap.txt ' dirNameAnalysisFollLinux];
+        end
         [statusCopyAnalysisFile1, cmdoutCopyAnalysisFile1] = system(commandCopyAnalysisFile1);
         disp(cmdoutCopyAnalysisFile1)
     end
     
     if(exist([fileNameFoll(1:end-4) '_obs_stats.txt'], 'file') == 2)
-        commandCopyAnalysisFile2 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_obs_stats.txt ' dirNameAnalysisFollLinux];
+        if(ispc)
+            commandCopyAnalysisFile2 = ['move ' dirNameDataFoll fileNameFoll(1:end-4) '_obs_stats.txt ' dirNameAnalysisFoll(1:end-1)];
+        else
+            commandCopyAnalysisFile2 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_obs_stats.txt ' dirNameAnalysisFollLinux];
+        end
         [statusCopyAnalysisFile2, cmdoutCopyAnalysisFile2] = system(commandCopyAnalysisFile2);
         disp(cmdoutCopyAnalysisFile2)
     end
     
     if(exist([fileNameFoll(1:end-4) '_posterior_s.txt'], 'file') == 2)
-        commandCopyAnalysisFile3 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_posterior_s.txt ' dirNameAnalysisFollLinux];
+        if(ispc)
+            commandCopyAnalysisFile3 = ['move ' dirNameDataFoll fileNameFoll(1:end-4) '_posterior_s.txt ' dirNameAnalysisFoll(1:end-1)];
+        else
+            commandCopyAnalysisFile3 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_posterior_s.txt ' dirNameAnalysisFollLinux];
+        end
         [statusCopyAnalysisFile3, cmdoutCopyAnalysisFile3] = system(commandCopyAnalysisFile3);
         disp(cmdoutCopyAnalysisFile3)
     end
     
     if(exist([fileNameFoll(1:end-4) '_posterior_N.txt'], 'file') == 2)
-        commandCopyAnalysisFile4 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_posterior_N.txt ' dirNameAnalysisFollLinux];
+        if(ispc)
+            commandCopyAnalysisFile4 = ['move ' dirNameDataFoll fileNameFoll(1:end-4) '_posterior_N.txt ' dirNameAnalysisFoll(1:end-1)];
+        else
+            commandCopyAnalysisFile4 = ['mv ' dirNameDataFollLinux fileNameFoll(1:end-4) '_posterior_N.txt ' dirNameAnalysisFollLinux];
+        end
         [statusCopyAnalysisFile4, cmdoutCopyAnalysisFile4] = system(commandCopyAnalysisFile4);
         disp(cmdoutCopyAnalysisFile4)
     end
     
     cd(mainDir)
 
+    % check if some sites were flipped, adjust sign to accomodate
     posterior_s = dlmread([dirNameAnalysisFoll fileNameFoll(1:end-4) '_posterior_s.txt']);
     temp2000 = repmat(flipVec', 1, size(posterior_s, 2));
     posterior_s = posterior_s.*temp2000;
@@ -168,19 +285,64 @@ for thisItr = 1:numItr
     end
 
     timeWholeCodeFoll(thisItr) = toc;
-    timeWholeCodeFoll(thisItr)
+    disp([' Time required to analyze this iteration: ' num2str(timeWholeCodeFoll(thisItr))])
         save([dirNameAnalysisFoll fileNameFoll(1:end-4) '_posterior_s.mat'])%, ...
             %'posterior_s', 'sigmaEstOutFoll_mean', 'sigmaEstOutFoll_median', 'sigmaEstOutFoll_ML', 'timeWholeCodeFoll')
     if(thisItr == numItr)
-        commandRemoveFile1 = ['rm ' destDirNameLinux wfabcFile1];
+        if(ispc)
+            commandRemoveFile1 = ['delete ' destDirNameLinux wfabcFile1 '.exe'];
+        else
+            commandRemoveFile1 = ['rm ' destDirNameLinux wfabcFile1];
+        end
+        
         [statusRemoveFile1, cmdoutRemoveFile1] = system(commandRemoveFile1);
-        commandRemoveFile2 = ['rm ' destDirNameLinux wfabcFile2];
+        if(ispc)
+            commandRemoveFile2 = ['delete ' destDirNameLinux wfabcFile2 '.exe'];
+        else
+            commandRemoveFile2 = ['rm ' destDirNameLinux wfabcFile2];
+        end
+        
         [statusRemoveFile2, cmdoutRemoveFile2] = system(commandRemoveFile2);
     end
+    
+    posOnlyItrTemp = perSiteSelction == max(perSiteSelction);
+    negOnlyItrTemp = perSiteSelction == min(perSiteSelction);
+    [~,~,~, aucFoll_MLItrTemp(1)] = perfcurve(double(posOnlyItrTemp), sigmaEstOutFoll_ML, 1);
+    [~,~,~, aucFoll_MLItrTemp(2)] = perfcurve(double(~negOnlyItrTemp), sigmaEstOutFoll_ML, 1);
+
+    if(thisItr == 1)
+        str = ' ';
+        str = [str ',trajectory,method,t0,T,ns,deltat,runtime'];
+        for l = 1:Lin
+            str = [str ',s' num2str(l-1)];
+        end
+        str = [str ',AUROC_ben,AUROC_del'];
+        for l = 1:Lin
+            str = [str ',ds' num2str(l-1)];
+        end
+        str = [str '\n'];
+        fid = fopen(fileNamePaperDataCompABC,'wt');
+        fprintf(fid, str);
+        fclose(fid);
+    end
+
+    inpDataABC = [num2str(thisItr-1) ',' num2str(thisItr-1) ',ABC,' num2str(Tstart) ',' num2str(T) ',' num2str(ng) ',' num2str(dT) ',' num2str(timeWholeCodeFoll(thisItr))];
+    for l = 1:Lin
+        inpDataABC = [inpDataABC ',' num2str(sigmaEstOutFoll_ML(l))];
+    end
+    inpDataABC = [inpDataABC ',' num2str(aucFoll_MLItrTemp(1)) ',' num2str(aucFoll_MLItrTemp(2))];
+    for l = 1:Lin
+        inpDataABC = [inpDataABC ',' num2str(perSiteSelction(l) - sigmaEstOutFoll_ML(l))];
+    end
+    inpDataABC = [inpDataABC '\n'];
+    
+    
+    fid = fopen([dirNameSaveCSVFile fileNamePaperDataCompABC],'a');
+    fprintf(fid, inpDataABC);
+    fclose(fid);
+    errorExit = 0;
 end
-%%
-disp('Time required to run the code')
-[timeToStartWFABC1;
-timeWFABC1;
-timeWFABC2;
-timeWholeCodeFoll]
+if(errorExit == 1)
+else
+    disp(['Output saved to ...' chosenSlash 'Matlab' chosenSlash  fileNamePaperDataCompABC])
+end
